Inject a muzzle into Bucket to remove knowledge of active record.
The goal is to decouple the logic of bucket's speaking ability away from
the knowledge of persisting that state to the database.

I think there could be a charter around checking out if Bucket is ever
told to shut up forever - there might be an opportunity to delete some
code!

Adding the clasp_repo contract to rubocop's ignore file because it's
treating a spec definition inside of a method as too much complication.
Which may be true, but IMO it's still readable.

[#78772282]